wrote test for procedures function in loadingFuncs.ts
