+++ package.json
@@ -19,6 +19,7 @@
     "date-fns": "^2.23.0",
     "firebase": "8.10.0",
     "firebase-tools": "^9.16.6",
+    "mockdate": "^3.0.5",
     "next": "11.0.1",
     "next-transpile-modules": "^8.0.0",
     "react": "17.0.2",
